Add http network retrieval api
Add http network retrieval API that returns the network map of the local
host.

Fix serialization.

To Use:

Navigate to the same port in the browser as you are running the health
check udp port on. The root route will return the network table of the
hosting server.

Subject to change.